Created files for the classes to tidy up the file more.
Json Files are being compared and differences are written in a separate dictionary
