+++ EpisodeData.json
@@ -1,18 +1,22 @@
 {
   "onePiece": {
-    "Episode": "838",
-    "Last Upload": "6 days ago"
+    "Episode": "847",
+    "Last Upload": "18 hrs ago"
   },
   "boruto": {
-    "Episode": "60",
-    "Last Upload": "1 day ago"
+    "Episode": "66",
+    "Last Upload": "3 days ago"
   },
   "foodWars": {
-    "Episode": "8",
-    "Last Upload": "5 days ago"
+    "Episode": "12",
+    "Last Upload": "35 days ago"
   },
   "boku": {
-    "Episode": "9",
-    "Last Upload": "44 mins ago"
+    "Episode": "16",
+    "Last Upload": "1 day ago"
+  },
+  "cellsAtWork": {
+    "Episode": "4",
+    "Last Upload": "1 day ago"
   }
 }
