+++ EpisodeData.py
@@ -1,0 +1,12 @@
+class EpisodeData():        
+    onePiece = {}
+    boruto = {}
+    foodWars = {}
+    boku = {}
+    cellsAtWork = {}
+    def __init__(self):
+        self.onePiece = {}
+        self.boruto = {}
+        self.foodWars = {}
+        self.boku = {}
+        self.cellsAtWork = {}
